// src/app/interfaces/respuesta-calculadora.ts
export interface RespuestaCalculadora {
    valorCuotaConSeguro: any;
    valorCuotaSinSeguro: number;
    valorTotalSeguro: number;
    costoMensualSeguro: number;
    montoTotalFinanciamiento: number;
    valorConDescuento: number;
    numeroCuotas: number;
    montoSolicitado: number;
    nominalMesVencido: any;
    seguroCuota: number;
    tasaEfectivaAnual: number;
    periodoGracia: number;
    cuatroPorMil: number;
    costoDeInteres: number;
    valorFuturo: number;
    valorFuturoSeguro: number;
    cuotaInicial: number;
    valorDescuento: number;
}